% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbours.R
\name{find_neighbours}
\alias{find_neighbours}
\title{Find the names of the neighbouring constituencies of a unit}
\usage{
find_neighbours(cont, x)
}
\arguments{
\item{cont}{A neighbourhood list of class "nb"}

\item{x}{The name of a unit}
}
\value{
A vector containing the names of its neighbours
}
\description{
Find the names of the neighbouring constituencies of a unit
}
