% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_output}
\alias{get_output}
\title{Title}
\usage{
get_output(model, dataframe)
}
\arguments{
\item{model}{The output of an `mgcv::gam()` model}

\item{dataframe}{`sf` dataframe to which the output will be attached}
}
\value{
`sf` dataframe with the smooth and fixed outputs of the model as the initial columns
}
\description{
Title
}
