% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_contigs.R
\name{make_cont_k_islands}
\alias{make_cont_k_islands}
\title{Title}
\usage{
make_cont_k_islands(data, unit, link_islands_k = 0)
}
\arguments{
\item{data}{A simple features dataframe}

\item{unit}{The level (e.g. region, county) at which the neighbourhood structure operates}

\item{link_islands_k}{An integer value. The closest k units to isolated units will be represented as neighbours}
}
\value{
A neighbourhood list of class "nb"
}
\description{
Title
}
