% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_link_numeric.R
\name{manual_link_numeric}
\alias{manual_link_numeric}
\title{Add an additional link to a neighbourhood list by index numbers}
\usage{
manual_link_numeric(cont, x, y)
}
\arguments{
\item{cont}{A neighbourhood list of class "nb"}

\item{x}{Name of one spatial unit to connect}

\item{y}{Name of other spatial unit to connect}
}
\value{
A neighbourhood list of class "nb" with selected units connected
}
\description{
Add an additional link to a neighbourhood list by index numbers
}
