% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_unlink_name.R
\name{manual_unlink_name}
\alias{manual_unlink_name}
\title{Remove a link from a neighbourhood list by name}
\usage{
manual_unlink_name(cont, x, y)
}
\arguments{
\item{cont}{A neighbourhood list of class "nb"}

\item{x}{Name of one spatial unit to disconnect}

\item{y}{Name of other spatial unit to disconnect}
}
\value{
A neighbourhood list of class "nb" with selected units disconnected
}
\description{
Remove a link from a neighbourhood list by name
}
