% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_unlink_numeric.R
\name{manual_unlink_numeric}
\alias{manual_unlink_numeric}
\title{Remove a link from a neighbourhood list by index numbers}
\usage{
manual_unlink_numeric(cont, x, y)
}
\arguments{
\item{cont}{A neighbourhood list of class "nb"}

\item{x}{Index number of one spatial unit to disconnect}

\item{y}{Index number of other spatial unit to disconnect}
}
\value{
A neighbourhood list of class "nb" with selected units disconnected
}
\description{
Remove a link from a neighbourhood list by index numbers
}
