% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmap.R
\name{quickmap}
\alias{quickmap}
\title{Generates quick maps of random and mrf effects from mgcv models}
\usage{
quickmap(output)
}
\arguments{
\item{output}{the `sf` dataframe output of `get_output()`}
}
\value{
a list containing quick maps of all smooths
}
\description{
Generates quick maps of random and mrf effects from mgcv models
}
