% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmap_contigs.R
\name{quickmap_contigs}
\alias{quickmap_contigs}
\title{Title}
\usage{
quickmap_contigs(cont, data, unit)
}
\arguments{
\item{cont}{A neighbourhood list of class "nb"}

\item{data}{A simple features dataframe}

\item{unit}{The level (e.g. region, county) at which the neighbourhood structure operates}
}
\value{
A map for visualising and checking of contiguities
}
\description{
Title
}
