% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmap_smooths.R
\name{quickmap_smooths}
\alias{quickmap_smooths}
\title{Title}
\usage{
quickmap_smooths(output)
}
\arguments{
\item{output}{the `sf` dataframe output of `get_output()`}
}
\value{
quick maps of all smooths
}
\description{
Title
}
