% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_estimates.R
\name{tidy_estimates}
\alias{tidy_estimates}
\title{Title}
\usage{
tidy_estimates(model, dataframe)
}
\arguments{
\item{model}{The output of an `mgcv::gam()` model}

\item{dataframe}{`sf` dataframe to which the output will be attached}
}
\value{
`sf` dataframe with the smooth and fixed outputs of the model as the initial columns
}
\description{
Title
}
